+++ docs/ops/cli.zh.md
@@ -100,40 +100,40 @@ available.
 
 -   提交一个Python Table的作业:
 
-        ./bin/flink run -py WordCount.py -j <path/to/flink-table.jar>
+        ./bin/flink run -py WordCount.py -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个有多个依赖的Python Table的作业:
 
-        ./bin/flink run -py examples/python/table/batch/word_count.py -j <path/to/flink-table.jar> \
+        ./bin/flink run -py examples/python/table/batch/word_count.py -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar> \
                                 -pyfs file:///user.txt,hdfs:///$namenode_address/username.txt
 
 -   提交一个有多个依赖的Python Table的作业，Python作业的主入口通过pym选项指定:
 
-        ./bin/flink run -pym batch.word_count -pyfs examples/python/table/batch -j <path/to/flink-table.jar>
+        ./bin/flink run -pym batch.word_count -pyfs examples/python/table/batch -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个指定并发度为16的Python Table的作业:
 
-        ./bin/flink run -p 16 -py examples/python/table/batch/word_count.py -j <path/to/flink-table.jar>
+        ./bin/flink run -p 16 -py examples/python/table/batch/word_count.py -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个关闭flink日志输出的Python Table的作业:
 
-        ./bin/flink run -q -py examples/python/table/batch/word_count.py -j <path/to/flink-table.jar>
+        ./bin/flink run -q -py examples/python/table/batch/word_count.py -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个运行在detached模式下的Python Table的作业:
 
-        ./bin/flink run -d -py examples/python/table/batch/word_count.py -j <path/to/flink-table.jar>
+        ./bin/flink run -d -py examples/python/table/batch/word_count.py -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个运行在指定JobManager上的Python Table的作业:
 
         ./bin/flink run -m myJMHost:8081 \
                             -py examples/python/table/batch/word_count.py \
-                            -j <path/to/flink-table.jar>
+                            -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
 
 -   提交一个运行在有两个TaskManager的[per-job YARN cluster]({{site.baseurl}}/ops/deployment/yarn_setup.html#run-a-single-flink-job-on-hadoop-yarn)的Python Table的作业:
 
         ./bin/flink run -m yarn-cluster -yn 2 \
                                  -py examples/python/table/batch/word_count.py \
-                                 -j <path/to/flink-table.jar>
+                                 -C <path/to/flink-table.jar> -C <path/to/flink-python-*-java-binding.jar>
                                  
 </div>
