+++ flink-clients/pom.xml
@@ -68,13 +68,6 @@ under the License.
 			<artifactId>commons-cli</artifactId>
 		</dependency>
 
-		<!-- Python API dependencies -->
-		<dependency>
-			<groupId>net.sf.py4j</groupId>
-			<artifactId>py4j</artifactId>
-			<version>${py4j.version}</version>
-		</dependency>
-
 		<!-- test dependencies -->
 
 		<dependency>
