+++ flink-dist/pom.xml
@@ -361,6 +361,14 @@ under the License.
 			<version>${project.version}</version>
 			<scope>provided</scope>
 		</dependency>
+
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-python</artifactId>
+			<classifier>java-binding</classifier>
+			<version>${project.version}</version>
+			<scope>provided</scope>
+		</dependency>
 		<!-- end optional Flink libraries -->
 
 		<!-- test dependencies -->
@@ -568,15 +576,6 @@ under the License.
 									<projectName>Apache Flink</projectName>
 								</transformer>
 							</transformers>
-							<relocations>
-								<relocation>
-									<pattern>py4j</pattern>
-									<shadedPattern>org.apache.flink.api.python.py4j</shadedPattern>
-									<includes>
-										<include>py4j.*</include>
-									</includes>
-								</relocation>
-							</relocations>
 						</configuration>
 					</execution>
 					<execution>
