+++ flink-dist/src/main/assemblies/opt.xml
@@ -184,6 +184,14 @@
 			<destName>flink-streaming-python_${scala.binary.version}-${project.version}.jar</destName>
 			<fileMode>0644</fileMode>
 		</file>
+
+		<!-- Python -->
+		<file>
+			<source>../flink-python/target/flink-python-${project.version}-java-binding.jar</source>
+			<outputDirectory>opt</outputDirectory>
+			<destName>flink-python-${project.version}-java-binding.jar</destName>
+			<fileMode>0644</fileMode>
+		</file>
 	</files>
 	<fileSets>
 		<fileSet>
