+++ flink-python/src/main/java/org/apache/flink/python/bridge/pickle/ArrayConstructor.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.python.pickle;
+package org.apache.flink.python.bridge.pickle;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
