+++ flink-python/src/main/java/org/apache/flink/python/client/PythonEnvUtils.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.client.python;
+package org.apache.flink.python.client;
 
 import org.apache.flink.core.fs.FileSystem;
 import org.apache.flink.core.fs.Path;
@@ -41,8 +41,8 @@ import java.util.UUID;
 /**
  * The util class help to prepare Python env and run the python process.
  */
-public final class PythonUtil {
-	private static final Logger LOG = LoggerFactory.getLogger(PythonUtil.class);
+public final class PythonEnvUtils {
+	private static final Logger LOG = LoggerFactory.getLogger(PythonEnvUtils.class);
 
 	private static final String FLINK_OPT_DIR = System.getenv("FLINK_OPT_DIR");
