+++ flink-python/src/test/java/org/apache/flink/python/client/PythonDriverTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.client.python;
+package org.apache.flink.python.client;
 
 import org.apache.flink.core.fs.Path;
